+++ craco.config.js
@@ -8,7 +8,8 @@ module.exports = {
       new CopyPlugin({
         patterns: [
           { from: 'node_modules/@hpcc-js/wasm/dist/graphvizlib.wasm', to: 'static/js' },
-          { from: 'node_modules/@hpcc-js/wasm/dist/expatlib.wasm', to: 'static/js' }
+          { from: 'node_modules/@hpcc-js/wasm/dist/expatlib.wasm', to: 'static/js' },
+          { from: 'node_modules/emojibase-data/en/data.json', to: 'static/js/emoji-data.json' }
         ],
       }),
       ...when(Boolean(process.env.ANALYZE), () => [