+++ src/components/editor/editor-pane/tool-bar/emoji-picker/emoji-picker-button.tsx
@@ -16,10 +16,13 @@ export const EmojiPickerButton: React.FC<EmojiPickerButtonProps> = ({ editor }) 
 
   return (
     <Fragment>
-      <EmojiPicker show={showEmojiPicker} onEmojiSelected={(emoji) => {
-        setShowEmojiPicker(false)
-        addEmoji(emoji, editor)
-      }} onDismiss={() => setShowEmojiPicker(false)}/>
+      <EmojiPicker
+        show={showEmojiPicker}
+        onEmojiSelected={(emoji) => {
+          setShowEmojiPicker(false)
+          addEmoji(emoji, editor)
+        }}
+        onDismiss={() => setShowEmojiPicker(false)}/>
       <Button variant='light' onClick={() => setShowEmojiPicker(old => !old)} title={t('editor.editorToolbar.emoji')}>
         <ForkAwesomeIcon icon="smile-o"/>
       </Button>