+++ src/components/editor/editor-pane/tool-bar/utils/emojiUtils.ts
@@ -1,15 +1,20 @@
-import { BaseEmoji, CustomEmoji, EmojiData } from 'emoji-mart'
+import { EmojiClickEventDetail, NativeEmoji } from 'emoji-picker-element/shared'
 
-export const getEmojiIcon = (emoji: EmojiData):string => {
-  if ((emoji as BaseEmoji).native) {
-    return (emoji as BaseEmoji).native
-  } else if ((emoji as CustomEmoji).imageUrl) {
+export const getEmojiIcon = (emoji: EmojiClickEventDetail): string => {
+  if (emoji.unicode) {
+    return emoji.unicode
+  }
+  if (emoji.name) {
     // noinspection CheckTagEmptyBody
-    return `<i class="fa ${(emoji as CustomEmoji).name}"></i>`
+    return `<i class="fa ${emoji.name}"></i>`
   }
   return ''
 }
 
-export const getEmojiShortCode = (emoji: EmojiData):string => {
-  return (emoji as BaseEmoji).colons
+export const getEmojiShortCode = (emoji: EmojiClickEventDetail): string => {
+  let skinToneModifier = ''
+  if ((emoji.emoji as NativeEmoji).skins && emoji.skinTone !== 0) {
+    skinToneModifier = `:skin-tone-${emoji.skinTone as number}:`
+  }
+  return `:${emoji.emoji.shortcodes[0]}:${skinToneModifier}`
 }